feat(examples/multiples_levels):  * moved level transition code to a seperate module  * added another level  * various minor tweaks  * fixed up camera logic some more
