+++ examples/bevy_gltf_blueprints/multiple_levels/assets/assets_game.assets.ron
@@ -1,6 +1,7 @@
 ({
     "world":File (path: "models/StartLevel.glb"),
     "level1":File (path: "models/Level1.glb"),
+    "level2":File (path: "models/Level2.glb"),
 
     "models": Folder (
         path: "models/library",
