+++ examples/bevy_gltf_blueprints/multiple_levels/src/assets/assets_game.rs
@@ -9,7 +9,8 @@ pub struct GameAssets {
     pub world: Handle<Gltf>,
     #[asset(key = "level1")]
     pub level1: Handle<Gltf>,
-
+    #[asset(key = "level2")]
+    pub level2: Handle<Gltf>,
     #[asset(key = "models", collection(typed, mapped))]
     pub models: HashMap<String, Handle<Gltf>>,
 }
