+++ examples/bevy_gltf_blueprints/multiple_levels/src/game/mod.rs
@@ -1,5 +1,4 @@
 pub mod in_game;
-use bevy_gltf_blueprints::GameWorldTag;
 pub use in_game::*;
 
 pub mod in_main_menu;
@@ -7,6 +6,9 @@ pub use in_main_menu::*;
 
 pub mod picking;
 pub use picking::*;
+
+pub mod level_transitions;
+pub use level_transitions::*;
 
 use crate::{
     insert_dependant_component,
@@ -88,86 +90,6 @@ pub fn test_collision_events(
     }
 }
 
-#[derive(Component, Reflect, Default, Debug)]
-#[reflect(Component)]
-pub struct LevelTransition{
-    pub target: String
-}
-
-pub fn trigger_level_transition(
-    mut collision_events: EventReader<CollisionEvent>,
-    level_transition_triggers: Query<&LevelTransition>,
-    parents: Query<&Parent>,
-
-    mut commands: Commands,
-    game_assets: Res<GameAssets>,
-    models: Res<Assets<bevy::gltf::Gltf>>,
-
-    game_world: Query<(Entity, &GameWorldTag)>
-){
-    for collision_event in collision_events.read() {
-        println!("collision");
-        match collision_event {
-            CollisionEvent::Started(entity1, entity2, _) => {
-                // we need to accomodate for the fact that the collider may be a child of the level transition (FIXME: is this a missunderstanding on my part about rapier child colliders ?)
-                let entity1_parent = parents.get(*entity1).unwrap();
-                let entity2_parent = parents.get(*entity2).unwrap();
-                if 
-                    level_transition_triggers.get(*entity1).is_ok() ||  
-                    level_transition_triggers.get(*entity2).is_ok() ||  
-                    level_transition_triggers.get(entity1_parent.get()).is_ok() || 
-                    level_transition_triggers.get(entity2_parent.get()).is_ok()
-                {
-                    println!("collision started, we can transition to level");
-                    let transition_trigger;
-                    if level_transition_triggers.get(*entity1).is_ok() {
-                        transition_trigger = level_transition_triggers.get(*entity1).unwrap();
-                    }else if  level_transition_triggers.get(*entity2).is_ok() {
-                        transition_trigger = level_transition_triggers.get(*entity2).unwrap();
-                    }else if level_transition_triggers.get(entity1_parent.get()).is_ok(){
-                        transition_trigger = level_transition_triggers.get(entity1_parent.get()).unwrap();
-                    }
-                    else {
-                        transition_trigger = level_transition_triggers.get(entity2_parent.get()).unwrap();
-                    }
-                    let current_game_world = game_world.single();
-
-                    // remove current level/world
-                    info!("despawning current level");
-                    commands.entity(current_game_world.0).despawn_recursive();
-
-                    let target_level = &transition_trigger.target;
-                    let level;
-                    println!("target level {}", target_level);
-                    if target_level == "Level1"{
-                        level = &game_assets.level1;
-                    }else {
-                        level = &game_assets.world;
-                    }
-                    info!("spawning new level");
-                    commands.spawn((
-                        SceneBundle {
-                            // note: because of this issue https://github.com/bevyengine/bevy/issues/10436, "world" is now a gltf file instead of a scene
-                            scene: models
-                                .get(level.id())
-                                .expect("main level should have been loaded")
-                                .scenes[0]
-                                .clone(),
-                            ..default()
-                        },
-                        bevy::prelude::Name::from("world"),
-                        GameWorldTag,
-                        InAppRunning,
-                    ));
-
-                }
-            }
-            CollisionEvent::Stopped(_entity1, _entity2, _) => {
-                println!("collision ended")
-            }
-        }
-    }
-}
 
 pub struct GamePlugin;
 impl Plugin for GamePlugin {
